add unicode to set, tighten bin

+++ core/utility/registry.h
@@ -156,7 +156,8 @@ namespace registry
 	void WriteStringToRegistry(_In_ HKEY hKey, _In_ const std::wstring& szValueName, _In_ const std::wstring& szValue);
 	void
 	WriteBinToRegistry(_In_ HKEY hKey, _In_ const std::wstring& szValueName, _In_ const std::vector<BYTE>& binValue);
-	inline void WriteBinToRegistry(_In_ HKEY hKey, _In_ const std::wstring& szValueName, _In_ ULONG cb, _In_ const BYTE* lpbin)
+	inline void
+	WriteBinToRegistry(_In_ HKEY hKey, _In_ const std::wstring& szValueName, _In_ size_t cb, _In_ const BYTE* lpbin)
 	{
 		WriteBinToRegistry(hKey, szValueName, std::vector<BYTE>(lpbin, lpbin + cb));
 	}
